Switched to using seconds instead of hours internally
This fixes multiple issues/bugs with wrong cheapest time interval lengths due to floating point numbers.

+++ App/AWattPrice/AWattPrice/Extensions/Formatter/TotalTimeFormatter.swift
@@ -8,7 +8,7 @@
 import Foundation
 
 class TotalTimeFormatter {
-    func localizedTotalTimeString(hour: Double, minute: Double) -> String {
+    func localizedTotalTimeString(hour: Int, minute: Int) -> String {
         let numberFormatter = NumberFormatter()
         numberFormatter.numberStyle = .none
         let hourString = numberFormatter.string(for: hour) ?? ""
